mindSum done.

+++ src/main/java/competativeCoding/octlong/Sweet.java
@@ -2,10 +2,10 @@ package competativeCoding.octlong;
 
 public class Sweet {
 
-    public int remem;
-    public int num;
+    public long remem;
+    public long num;
 
-    public Sweet(int num, int remem) {
+    public Sweet(long num, long remem) {
         this.remem = remem;
         this.num = num;
     }
